% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aminoAcids}
\alias{aminoAcids}
\title{Amino Acids and their Basepair representation}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
aminoAcids
}
\description{
Shows all the possible basepair triplets related to all the aminoacids.
}
\keyword{datasets}
