% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ander_function.R
\name{ander_1}
\alias{ander_1}
\title{Ander_function}
\usage{
ander_1(ander_2)
}
\arguments{
\item{ander_2}{here it goes the description of the parameter}
}
\value{
here it goes the description of function output
}
\description{
Ander_function
}
