% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon.R
\name{codon}
\alias{codon}
\title{Separate to codons function}
\usage{
codon(sequence, start = 1)
}
\arguments{
\item{sequence}{The nucleotide sequence string}

\item{start}{Position of where to start looking for start codon}
}
\value{
The function returns the start and end codon of a nucletotide sequence
}
\description{
Separate to codons function
}
\examples{
codon("AUGGCUUAA")
codon("ATTGCA", start=3)
codon("ATG")
codon("AT")    # returns ""
}
