% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to_amino_acids.R
\name{codon_to_amino_acids}
\alias{codon_to_amino_acids}
\title{Convert codon string into amno acids}
\usage{
codon_to_amino_acids(codons)
}
\arguments{
\item{codons}{Sequence of codons to encode into amino acids}
}
\value{
The codon string translated into amino acids
}
\description{
Convert codon string into amno acids
}
\examples{
codon_to_amino_acids(c("AUG", "UUU", "GGC"))
}
