% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to_name.R
\name{codon_to_name}
\alias{codon_to_name}
\title{Codon to name}
\usage{
codon_to_name(codons)
}
\arguments{
\item{codons}{A character vector of codon codes (e.g., "AUG", "UUU") to be translated}
}
\value{
A single character string with the concatenated one letter codes for the codons
}
\description{
Codon to name
}
\examples{
codon_to_name(c("AUG", "UUU", "UUA"))
}
