% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_sample.R
\name{DNA_sample}
\alias{DNA_sample}
\title{DNA Sample}
\usage{
DNA_sample(amount_nucleotides)
}
\arguments{
\item{amount_nucleotides}{the amount of nucleotides in a given DNA sequence}
}
\value{
The function output is a random DNA sequence consisting of the bases A, T, G and C. The amount of bases in the sequence is chosen by the user.
}
\description{
DNA Sample
}
\examples{
amount_nucleotides <- 10
DNA_sample(amount_nucleotides)
}
