% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{Converting a DNA sequnce into a RNA sequence}
\usage{
DNA_to_RNA(DNA1)
}
\arguments{
\item{DNA1}{The DNA sequnce that we would like to convert into RNA.}
}
\value{
The output of this function is that our DNA sequcne have been converted into a RNA sequence.
}
\description{
Converting a DNA sequnce into a RNA sequence
}
\examples{
DNA1 <- "ATGCTATGA"
DNA_to_RNA(DNA1)
}
