% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna_sequence.R
\name{generate_dna_sequence}
\alias{generate_dna_sequence}
\title{Generate a random DNA sequence}
\usage{
generate_dna_sequence(sequence_length)
}
\arguments{
\item{sequence_length}{Integer specifying the length of the DNA sequence to generate.}
}
\value{
A string representing the randomly generated DNA sequence using A, T, G, and C nucleotides.
}
\description{
The function generates a random DNA sequence of a specified length imputed and
using the nucleotides A, T, G, and C.
}
\examples{
# Generate a sequence of length 10 with A, T, G, and C nucleotides.
generate_dna_sequence(10)

# Generate a sequence of length 50 with A, T, G, and C nucleotides.
generate_dna_sequence(50)
}
