% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_codon.R
\name{make_codon}
\alias{make_codon}
\title{Converting nucleotides into codons}
\usage{
make_codon(nucleotides, start = 1)
}
\arguments{
\item{nucleotides}{All nucleotides}

\item{start}{The first nucleotide}
}
\value{
Nucleotides divided into codons
}
\description{
Converting nucleotides into codons
}
\examples{
nucleotides <- "AUGCUU"
make_codon(nucleotides)
}
