% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amino_acid_counts.R
\name{plot_amino_acid_counts}
\alias{plot_amino_acid_counts}
\title{Plot Amino Acid Counts in a Sequence}
\usage{
plot_amino_acid_counts(sequence)
}
\arguments{
\item{sequence}{A character string representing a sequence of amino acids}
}
\value{
A ggplot object displaying a bar plot of amino acid counts
}
\description{
Plot Amino Acid Counts in a Sequence
}
\examples{
sequence <- "AAGTCCGAA"
plot_amino_acid_counts(sequence)
}
