% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_char_counts.R
\name{plot_char_counts}
\alias{plot_char_counts}
\title{Plot Counts per Unique Character}
\usage{
plot_char_counts(input)
}
\arguments{
\item{input}{The input string we want the unique character counts of.}
}
\value{
A plot of of the counts per unique character contained within the string.
}
\description{
Plot Counts per Unique Character
}
