% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDNAseq.R
\name{randomDNAseq}
\alias{randomDNAseq}
\title{randomDNAseq}
\usage{
randomDNAseq(len)
}
\arguments{
\item{len}{lenght of the sequence}
}
\value{
random sample of DNAseq of the length we want
}
\description{
randomDNAseq
}
