% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Title}
\usage{
transcription(data)
}
\arguments{
\item{data}{}
}
\value{
RNA sequence from initial DNA data
}
\description{
Title
}
\examples{
dna_seq <- c("A", "G", "A", "T", "T", "C", "T", "G", "C", "C", "T")
rna_seq <- transription(dna_seq)
}
