% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RN.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{Convert a DNA sequence string into an RNA sequence string}
\usage{
DNA_to_RNA(dna_sequence)
}
\arguments{
\item{dna_sequence}{A string representing a DNA sequence using A, T, G, and C nucleotides.}
}
\value{
A string representing the equivalent RNA sequence with U nucleotides.
}
\description{
Given a DNA sequence string, this function returns the corresponding RNA sequence
by replacing all T nucleotides with U.
}
\examples{
DNA_to_RNA("ATGC")   # returns "AUGC"
}
